Fixed a few more bugs to get the Multi-core version to work
Also changed the Reduce pattern to return long on getIteration().

+++ PGAFramework/src/edu/uncc/grid/pgaf/datamodules/ReduceDataState.java
@@ -4,5 +4,5 @@ import java.io.Serializable;
 
 public interface ReduceDataState extends Serializable {
 	public void advanceIteration();
-	public int getIteration();
+	public long getIteration();
 }
